Correct command tx/rx location, add 242 endpoint with clusters.
