+++ zigpy/zcl/clusters/__init__.py
@@ -5,6 +5,7 @@ import inspect
 from . import (
     closures,
     general,
+    greenpower,
     homeautomation,
     hvac,
     lighting,
@@ -23,6 +24,7 @@ CLUSTERS_BY_NAME: dict[str, Cluster] = {}
 for cls in (
     closures,
     general,
+    greenpower,
     homeautomation,
     hvac,
     lighting,
